pyproject.toml
[build-system]
requires = ["maturin>=1.2,<2.0"]
build-backend = "maturin"

[project]
name = "letsql"
version = "0.1.2"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
    "ibis-framework==7.2.0"
]


[tool.maturin]
module-name = "letsql._internal"
python-source = "python"
features = ["pyo3/extension-module"]

[tool.codespell]
skip = "*.lock,.direnv,.git,docs/_freeze/**/html.json"
ignore-regex = '\b(DOUB|i[if]f|I[IF]F|lamduh|AFE|crate|ba)\b'
builtin = "clear,rare,names"

[tool.poetry]
name = "letsql"
version = "0.1.2"
description = ""
authors = ["Hussain Sultan <hussain@letsql.com>", "Dan Lovell <dan@letsql.com>", "Daniel Mesejo <mesejo@letsql.com>"]
packages = [
	{ include = "letsql", from = "python" },
]

[tool.poetry.dependencies]
python = "^3.10"
maturin = "^1.4.0"
ibis-framework = "^7.2.0"
dask = "^2023.12.1"
xgboost = ">=1.6.1"
attrs = "^23.2.0"
connectorx = "^0.3.2"
scikit-learn = "^1.4.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"

[tool.poetry.group.dev.dependencies]
ipython = "^8.19.0"
pytest-cov = "4.1.0"
pytest = "7.4.2"
coverage = {version = "7.3.2", extras = ["toml"]}
pre-commit = "3.5.0"
ruff = "0.1.1"
black = "23.10.1"
blackdoc = "0.3.8"
codespell = {version = "2.2.6", extras = ["hard-encoding-detection", "toml"]}
pyarrow = "13.0.0"
datafusion = "^34.0.0"

